unnecessary casts.
git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk/htmlunit@14662 5f5364db-9458-4db8-a492-e30667be6df6

+++ src/main/java/com/gargoylesoftware/htmlunit/javascript/host/ArrayCustom.java
@@ -52,7 +52,7 @@ public final class ArrayCustom {
             final Object length = scriptable.get("length", scriptable);
             if (length != Scriptable.NOT_FOUND) {
                 final int size = (int) Context.toNumber(length);
-                array = new Object[(int) size];
+                array = new Object[size];
                 for (int i = 0; i < size; i++) {
                     array[i] = scriptable.get(i, scriptable);
                 }
